requirements.txt
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.3.6
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.109.2
fastapi-pagination==0.13.1
flower==2.0.1
greenlet==3.2.2
h11==0.16.0
humanize==4.12.3
idna==3.10
kombu==5.5.3
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
prometheus-client==0.19.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.2
redis==5.0.1
rsa==4.9.1
ruff==0.11.10
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tornado==6.4
typing_extensions==4.13.2
tzdata==2025.2
uv==0.7.3
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
fastapi-pagination==0.13.1
colorlog==6.9.0